% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviz_plot.R
\name{reviz_plot}
\alias{reviz_plot}
\title{Visualize relationships with model overlay}
\usage{
reviz_plot(data, x, y, z = NULL, model = c("lm", "loess"), use_color = TRUE)
}
\arguments{
\item{data}{A data.frame}

\item{x}{X variable name}

\item{y}{Y variable name}

\item{z}{Optional grouping variable for color/facets}

\item{model}{"lm" or "loess"}

\item{use_color}{TRUE for color, FALSE for facets}
}
\value{
ggplot object
}
\description{
Visualize relationships with model overlay
}
